+++ tools/rescue/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod admin_script_builder;
-pub mod rescue_cli;
+pub mod mission;
 pub mod framework_payload;
+pub mod diem_db_bootstrapper;
