+++ tools/rescue/src/main.rs
@@ -1,6 +1,6 @@
 
 use clap::{Parser, Subcommand};
-use rescue::rescue_cli::MissionOpts;
+use rescue::{mission::RescueTxOpts, diem_db_bootstrapper::BootstrapOpts};
 
 #[derive(Parser)]
 #[clap(author, version, about, long_about = None)]
@@ -12,15 +12,22 @@ struct RescueCli {
 
 #[derive(Subcommand)]
 enum Sub {
-    Mission(MissionOpts),
+    RescueTx(RescueTxOpts),
+    Bootstrap(BootstrapOpts)
 }
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
     let cli = RescueCli::parse();
     match cli.command {
-        Some(Sub::Mission(mission)) => {
+        Some(Sub::RescueTx(mission)) => {
             mission.run().await?;
+            println!("SUCCESS: rescue mission complete.")
+
+        },
+        Some(Sub::Bootstrap(bootstrap)) => {
+            bootstrap.run()?;
+            println!("SUCCESS: db boostrapped with writeset (genesis tx)")
         }
         _ => {
             println!("\nI'll be there")
